requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
motor==3.7.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
httpx==0.28.1
python-jose[cryptography]==3.5.0
cryptography==42.0.7
passlib[bcrypt]==1.7.4
redis==5.1.0
celery==5.3.6
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
pymongo==4.10.1
dnspython==2.5.0
pydantic[email]==2.11.7
email-validator
Pillow==10.4.0
requests==2.32.3
